+++ changelog.md
@@ -1,3 +1,11 @@
+2009.4.51
+-----------
+
+### Fix
+
+* translate lowercase header to normal header for env
+* don't get server name, that cause a crash
+
 2009.4.50
 ---------